+++ packages/terra-table/src/subcomponents/Cell.jsx
@@ -15,6 +15,7 @@ import VisuallyHiddenText from 'terra-visually-hidden-text';
 
 import ColumnContext from '../utils/ColumnContext';
 import GridContext, { GridConstants } from '../utils/GridContext';
+import getFocusableElements from '../utils/focusManagement';
 import styles from './Cell.module.scss';
 
 const cx = classNames.bind(styles);
@@ -149,12 +150,7 @@ function Cell(props) {
    * Determine if cell has focusable elements
    */
   const hasFocusableElements = () => {
-    const focusableElementSelector = "a[href]:not([tabindex='-1']), area[href]:not([tabindex='-1']), input:not([disabled]):not([tabindex='-1']), "
-      + "select:not([disabled]):not([tabindex='-1']), textarea:not([disabled]):not([tabindex='-1']), button:not([disabled]):not([tabindex='-1']), "
-      + "iframe:not([tabindex='-1']), [tabindex]:not([tabindex='-1']), [contentEditable=true]:not([tabindex='-1'])";
-
-    const focusableElements = [...cellRef.current.querySelectorAll(`${focusableElementSelector}`)].filter(el => !el.hasAttribute('disabled') && !el.getAttribute('aria-hidden'));
-
+    const focusableElements = getFocusableElements(cellRef.current);
     return focusableElements.length > 0;
   };