+++ packages/terra-table/CHANGELOG.md
@@ -6,6 +6,7 @@
   * Renamed `hasColumnHeaders` prop to `hasVisibleColumnHeaders`.
 
 * Added
+  * Added navigation via Home and End Keys.
   * Added title attribute to column header cells to enable tooltips.
 
 * Fixed